Core Data added for Downloads

+++ Swift-Netflix/Swift-Netflix/Models/Title.swift
@@ -12,13 +12,13 @@ struct TrendingTitleResponce: Codable {
 }
 
 struct Title: Codable {
-    let id: Int
+    let id: Int64
     let media_type: String?
     let original_name: String?
     let original_title: String?
     let poster_path: String?
     let overview: String?
-    let vote_count: Int
+    let vote_count: Int64
     let release_date: String?
     let vote_average: Double
 }
